Actualizacion de datos para requerimientos de TP (resubido desde PC)
subí los archivos desde un repo que no tenía actualizado, asi que lo resubí desde los archivos locales

+++ app/models/pagos.py
@@ -9,6 +9,6 @@ class Pagos(db.Model):
 
     id: int = db.Column('id', db.Integer, primary_key=True, autoincrement=True)
     producto_id: int = db.Column('producto_id', db.Integer, nullable=False)
-    precio: int = db.Column('precio', db.Integer, nullable=False)
+    precio: int = db.Column('precio', db.Float, nullable=False)
     medio_pago: str = db.Column('medio_pago', db.Text, nullable=False)
